refactor(editor): Extract per-thread runtime from Instance AI store (no-changelog)
Move per-thread state (messages, SSE connection, reducer routing, hydration)
and the methods that operate on it into a new `instanceAi.threadRuntime.ts`
module. The store keeps a single mutable runtime via `switchTo(threadId)`
and re-exports its refs/methods, so the store's external surface and all
23 consumer files are unchanged.

Also extracts two pure helpers from the runtime: `collapseDeltaEvents`
(text-delta/reasoning-delta merging in `copyFullTrace`) and
`buildRoutingFromMessages` (reducer routing rebuild during hydration).

Sets up the structural separation needed for future per-thread runtimes
(Map<threadId, ThreadRuntime>) without introducing them yet.
